+++ firstCarSogang_backend/firstCarSogang_tickets/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class FirstcarsogangTicketsConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'firstCarSogang_tickets'
