+++ firstCarSogang_backend/firstCarSogang_tickets/views.py
@@ -1,0 +1,4 @@
+from django.shortcuts import render
+
+def slowtrain(request):
+
